wp-automation/requirements.txt
# WordPress REST API
requests>=2.31.0

# AI API
openai>=1.0.0
anthropic>=0.18.0

# Template
jinja2>=3.1.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Utilities
rich>=13.0.0  # Pretty console output
